+++ version_history.md
@@ -1,5 +1,18 @@
 #Version history#
 
+##1.1 (08/05/12)
+* Added <tt>_.escape</tt> and <tt>_.esc</tt> alias
+* Added <tt>_.substitute</tt> and <tt>_.subst</tt> alias
+* Added <tt>_.isLuaKeyword</tt> and <tt>_.isLuaKword, _.isReserved</tt> aliases
+* Added <tt>_.isToken</tt> and <tt>_.isOperator, _.isOp</tt> aliases
+* Added <tt>_.isIdentifier</tt> and <tt>_.isIden, _.isName</tt> aliases
+* Added <tt>_.statistics</tt> and <tt>_.stats</tt> alias
+* Added <tt>_.byteAt</tt>
+* Added <tt>_.trim</tt>
+* Added string callable feature
+* Updated documentation
+* Updated Allen_Lib_Test.lua and Allen_StringOp_Tests.lua
+
 ##1.0 (08/02/12)
 * Added <tt>_.isAlpha</tt>, <tt>_.isNumeric</tt>, <tt>_.isAlphaNumeric</tt> and their aliases
 * Added <tt>_.isHexadecimal</tt> and <tt>_.isHex</tt> alias